Added service project, added eoc and drug to model, added repositories to support drugs and eocs.

+++ RemsLogic.Repositories/IRepository.cs
@@ -10,7 +10,7 @@ namespace RemsLogic.Repositories
     public interface IRepository<TModel>
         where TModel : IEntity
     {
-        TModel Get(Int32 id);
+        TModel Get(long id);
 
         void Save(TModel model);
         void Delete(TModel model);
